Remove unnecessary casts

+++ src/animation/AnimationRender.cpp
@@ -249,7 +249,7 @@ float Cedric_GetInvisibility(Entity *io) {
 			if(spell) {
 				if(player.m_skillFull.intuition > spell->m_level * 10) {
 					invisibility -= player.m_skillFull.intuition * (1.0f / 100)
-									+ (float)spell->m_level * (1.0f / 10);
+									+ spell->m_level * (1.0f / 10);
 
 					invisibility = glm::clamp(invisibility, 0.1f, 1.f);
 				}
